port: исправить гарантированные несовместимости с API Bannerlord 1.4.8
- SiegeWeapons: Mission.SpawnMissile(5-арг) -> AddExplosion (нет такой перегрузки)
- Weather: SetFog(..., 0x...) -> SetFog(..., (uint)0x...) (2-й аргумент uint)
- Components: OnTickAsAI -> OnTick в 4 классах (не виртуал AgentComponent)
- RoleComponents/FortressBuildManager: destructible.SetHitPoints -> HitPoints =
- WaveManager: убрано entryPoint.Direction (нет свойства у GameEntity)
- PropSpawner: MoveToFrame(Frame) -> MoveToFrame(Frame.ToMatrixFrame())
- docs/COMPILE_RISKS: каталог оставшихся неоднозначных API (MP-слой, AddExplosion,
  AddParticleSystem, weather setters, SetTargetForAI, campaign events)
- PROGRESS: запись о портировании

+++ BannerlordModule/src/NordInvasion/Behaviors/NordInvasionWeatherBehavior.cs
@@ -22,7 +22,7 @@ namespace NordInvasion.Behaviors
             switch (weather)
             {
                 case 1:
-                    Mission.Current.Scene.SetFog(30f, 0x888888);
+                    Mission.Current.Scene.SetFog(30f, (uint)0x888888);
                     InformationManager.DisplayMessage(new InformationMessage("Fog! Archers -50% range", Colors.Yellow));
                     break;
                 case 2:
@@ -39,7 +39,7 @@ namespace NordInvasion.Behaviors
                     InformationManager.DisplayMessage(new InformationMessage("Night! Torches needed!", Colors.Gold));
                     break;
                 default:
-                    Mission.Current.Scene.SetFog(100f, 0xFFFFFF);
+                    Mission.Current.Scene.SetFog(100f, (uint)0xFFFFFF);
                     Mission.Current.Scene.SetRainDensity(0f);
                     break;
             }
